Call To Action: Class migration done.
