+++ classes/class-uagb-block-helper.php
@@ -1240,9 +1240,11 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 			);
 
 			// @codingStandardsIgnoreEnd.
-			$desktop = UAGB_Helper::generate_css( $selectors, '#uagb-cta-block-' . $id );
-			$tablet  = UAGB_Helper::generate_css( $t_selectors, '#uagb-cta-block-' . $id );
-			$mobile  = UAGB_Helper::generate_css( $m_selectors, '#uagb-cta-block-' . $id );
+
+			$base_selector = ( $attr['classMigrate'] ) ? '.uagb-block-' : '#uagb-cta-block-';
+			$desktop = UAGB_Helper::generate_css( $selectors, $base_selector . $id );
+			$tablet  = UAGB_Helper::generate_css( $t_selectors, $base_selector . $id );
+			$mobile  = UAGB_Helper::generate_css( $m_selectors, $base_selector . $id );
 
 			$generated_css = array(
 				'desktop' => $desktop,
